remove warning ft_printf

+++ libft/srcs/ft_printf/tools_float.c
@@ -90,7 +90,7 @@ char	*ft_apply_round(char (*dst)[BUF_S], int stock, int precision)
 		if (!(tmp = apply_round_ni(*dst)))
 			return (NULL);
 	}
-	i = (i == -1) ? ft_strlen(*dst) - 1 : i - 1;
+	i = (i == -1) ? (int)ft_strlen(*dst) - 1 : i - 1;
 	stock = (stock > 4) ? 1 : 0;
 	apply_round_final(stock, &tmp, i);
 	return (tmp);
